// libs/quant-api/models/RetStockChaoDie.ts
/**
 * title
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { StockChaoDieJoda } from './StockChaoDieJoda';
import { HttpFile } from '../http/http';

export class RetStockChaoDie {
    'records'?: Array<StockChaoDieJoda>;
    'code': number;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "records",
            "baseName": "records",
            "type": "Array<StockChaoDieJoda>",
            "format": ""
        },
        {
            "name": "code",
            "baseName": "code",
            "type": "number",
            "format": "int32"
        }    ];

    static getAttributeTypeMap() {
        return RetStockChaoDie.attributeTypeMap;
    }

    public constructor() {
    }
}